a0: Reorganized, Notification
Reorganized the structure of the app. Created a service that utilizes a
custom LocationListener to hook into the location services of the phone.

Notification implementation is completed. Next task is to create a
browser intent to wiki (then specifically to the user's location).

*Broke main screen textview text in process of reorganization, need to
implement that after other key parts are finished.