alpha v1.00
App is functioning. Reorganized, turns out my initial scheme was better
for my purposes (or so I think right now). Notification now leads to the
wikipedia of the detected location.

Most pressing task is to figure out the lifecycle of the app when it is
in the background and the updating mechanics of the LocationListener in
this situation. More testing but very happy with progress!