Big Update: v1.1
Have worked out issues with the lifecycle of the app. Uses binding to
call methods from the custom LocationListener that is located in the
LocService created in the main activity. Now allows for the service to
remain alive and updating and also communicate with the main screen.

Added an icon for the app to use.

Need to implement the settings button.

App is functioning per plan. Need to do more testing in real life on my
S3, but initial results are exciting!